Fix tslint issues
