+++ src/components/MediocreMediaPlayerCard/components/VolumeSlider.tsx
@@ -15,7 +15,7 @@ const VolumeContainer = styled.div`
   margin-top: auto;
 `;
 
-const ControlButton = styled(IconButton)<{ muted?: boolean }>`
+const ControlButton = styled(IconButton) <{ muted?: boolean }>`
   opacity: ${props => (props.muted ? 0.8 : 1)}; // reduce opacity if muted
 `;
 
@@ -80,7 +80,7 @@ export const VolumeTrigger = ({
   setSliderVisible,
 }: {
   sliderVisible: boolean;
-  setSliderVisible: (boolean) => void;
+  setSliderVisible: (newValue: boolean) => void;
 }) => {
   const { hass, config } =
     useContext<CardContextType<MediocreMediaPlayerCardConfig>>(CardContext);
