+++ src/utils/GlanceGuard.tsx
@@ -1,6 +1,7 @@
+import { FC, JSX } from "preact/compat";
 import { useEffect, useState } from "preact/hooks";
 
-export const GlanceGuard: React.FC = ({ children }) => {
+export const GlanceGuard: FC<{children: JSX.Element}> = ({ children }) => {
   const [isGlanceLoaded, setIsGlanceLoaded] = useState(false);
   useEffect(() => {
     // https://community.home-assistant.io/t/re-using-existing-frontend-components-in-lovelace-card-editor/103415/9
